add end to end compiler tests
